Convert from command line args to ServerConfig yaml file
As part of #31 need to provide a bunch of database parameters
so it seemed like a good time to move to the configuration file.

Using yaml instead of json because it's much more human friendly,
will keep the world files as json (for now) because it's expected
that those will be written by some sort of builder tool and not
edited by hand.